Add files via upload
To understand session first we need to understand HTTP protocol .HTTP protocol is stateless because it send new request everytime when the user do any task in application . It doesn't remember which user request's lasttime so to make it remember that which user request lasttime(statefull) we use session and cookies . Session is store on server .Session store data temporarily. Every session has local ID .Session store this ID on the server and it also give this ID to cookie. Now when a user make any request the session first match it's ID with cookie ID. No whenever the will make any request the server will remember that which is requesting just because of session and cookies. Example when we  visit a shopping website after sometime it will recommend us product which is on the basis of our previous search this is because session ID when match with cookie ID the website recognize which user is requesting